+++ packages/xchain-thorchain-query/__e2e__/thorchain-liquidityP-e2e.ts
@@ -19,7 +19,7 @@ import {
   EstimateAddLP,
   EstimateAddSaver,
   EstimateWithdrawLP,
-  EstimateWithdrawSaver,
+  // EstimateWithdrawSaver,
   LiquidityPosition,
   SaversPosition,
   WithdrawLiquidityPosition,
@@ -35,7 +35,7 @@ const thorchainQuery = new ThorchainQuery(thorchainCache)
 
 // mainnet asset
 const BUSD = assetFromStringEx('BNB.BUSD-BD1')
-const synthBTC = assetFromStringEx('BTC/BTC')
+//const synthBTC = assetFromStringEx('BTC/BTC')
 const USDC = assetFromStringEx('ETH.USDC-0XA0B86991C6218B36C1D19D4A2E9EB0CE3606EB48')
 
 function printAdd(estimate: EstimateAddLP) {
@@ -96,9 +96,9 @@ function printSaver(saver: EstimateAddSaver) {
   const expanded = {
     assetAmount: saver.assetAmount.formatedAssetString(),
     fee: {
-      networkFee: saver.fee.networkFee.formatedAssetString(),
-      liquidityFee: saver.fee.liquidityFee.formatedAssetString(),
-      totalFees: saver.fee.totalFees.formatedAssetString(),
+      affiliateFee: saver.fee.affiliate.formatedAssetString(),
+      asset: saver.fee.asset.formatedAssetString(),
+      outbound: saver.fee.outbound.formatedAssetString(),
     },
     expiry: saver.expiry,
     toAddress: saver.toAddress,
@@ -107,21 +107,21 @@ function printSaver(saver: EstimateAddSaver) {
   }
   console.log(expanded)
 }
-function printWithdrawSaver(saver: EstimateWithdrawSaver) {
-  const expanded = {
-    assetAmount: saver.assetAmount.formatedAssetString(),
-    fee: {
-      networkFee: saver.fee.networkFee.formatedAssetString(),
-      liquidityFee: saver.fee.liquidityFee.formatedAssetString(),
-      totalFees: saver.fee.totalFees.formatedAssetString(),
-    },
-    expiry: saver.expiry,
-    toAddress: saver.toAddress,
-    memo: saver.memo,
-    estimateWaitTime: saver.estimatedWaitTime,
-  }
-  console.log(expanded)
-}
+// function printWithdrawSaver(saver: EstimateWithdrawSaver) {
+//   const expanded = {
+//     assetAmount: saver.assetAmount.formatedAssetString(),
+//     fee: {
+//       affili: saver.fee.affiliate.formatedAssetString(),
+//       liquidityFee: saver.fee.asset.formatedAssetString(),
+//       totalFees: saver.fee.outbound.formatedAssetString(),
+//     },
+//     expiry: saver.expiry,
+//     toAddress: saver.toAddress,
+//     memo: saver.memo,
+//     estimateWaitTime: saver.estimatedWaitTime,
+//   }
+//   console.log(expanded)
+// }
 function printSaversPosition(saver: SaversPosition) {
   const expanded = {
     depositValue: saver.depositValue.formatedAssetString(),
@@ -264,16 +264,11 @@ describe('Thorchain-amm liquidity action end to end Tests', () => {
     const estimateAddsSaver = await thorchainQuery.estimateAddSaver(addAssetAmount)
     printSaver(estimateAddsSaver)
   })
-  it(`Should estimate saver addition synth`, async () => {
-    const addAssetAmount = new CryptoAmount(assetToBase(assetAmount(1, 8)), synthBTC)
-    const estimateAddsSaver = await thorchainQuery.estimateAddSaver(addAssetAmount)
-    printSaver(estimateAddsSaver)
-  })
-  it(`Should estimate saver withdrawal`, async () => {
-    const addAssetAmount = new CryptoAmount(assetToBase(assetAmount(0, 8)), AssetBTC)
-    const estimateWithdrawSaver = await thorchainQuery.estimateWithdrawSaver(addAssetAmount)
-    printWithdrawSaver(estimateWithdrawSaver)
-  })
+  // it(`Should estimate saver withdrawal`, async () => {
+  //   const addAssetAmount = new CryptoAmount(assetToBase(assetAmount(0, 8)), AssetBTC)
+  //   const estimateWithdrawSaver = await thorchainQuery.estimateWithdrawSaver(addAssetAmount)
+  //   printWithdrawSaver(estimateWithdrawSaver)
+  // })
   it(`Should get saver position`, async () => {
     const address = 'bnb1jy7h4muz0ruflmrlxnt89ayhn3mf82sgza4vfm'
     const saver: getSaver = {