+++ packages/xchain-thorchain-query/src/utils/thornode.ts
@@ -10,6 +10,10 @@ import {
   Pool,
   PoolsApi,
   QueueApi,
+  QuoteApi,
+  QuoteSaverDepositResponse,
+  QuoteSaverWithdrawResponse,
+  QuoteSwapResponse,
   SaversApi,
   TransactionsApi,
   TxOutItem,
@@ -51,6 +55,7 @@ export class Thornode {
   private poolsApi: PoolsApi[]
   private liquidityProvidersApi: LiquidityProvidersApi[]
   private saversApi: SaversApi[]
+  private quoteApi: QuoteApi[]
 
   constructor(network: Network = Network.Mainnet, config?: ThornodeConfig) {
     this.network = network
@@ -66,6 +71,7 @@ export class Thornode {
       (url) => new LiquidityProvidersApi(new Configuration({ basePath: url })),
     )
     this.saversApi = this.config.thornodeBaseUrls.map((url) => new SaversApi(new Configuration({ basePath: url })))
+    this.quoteApi = this.config.thornodeBaseUrls.map((url) => new QuoteApi(new Configuration({ basePath: url })))
   }
 
   /**
@@ -202,6 +208,85 @@ export class Thornode {
     for (const api of this.saversApi) {
       try {
         const resp = (await api.saver(asset, address, height)).data
+        return resp
+      } catch (e) {
+        //console.error(e)
+      }
+    }
+    throw new Error(`THORNode not responding`)
+  }
+
+  /**
+   *
+   * @param asset - asset to add to savers
+   * @param amount - amount to deposit
+   * @param height - block height
+   * @returns quotes deposit object response
+   */
+  async getSaversDepositQuote(asset: string, amount: number, height?: number): Promise<QuoteSaverDepositResponse> {
+    for (const api of this.quoteApi) {
+      try {
+        const resp = (await api.quotesaverdeposit(height, asset, amount)).data
+        return resp
+      } catch (e) {
+        //console.error(e)
+      }
+    }
+    throw new Error(`THORNode not responding`)
+  }
+
+  /**
+   *
+   * @param asset - asset to withdraw
+   * @param address - savers address
+   * @param height - block height
+   * @param withdrawBps - withddraw percent
+   * @returns quotes withdraw object response
+   */
+  async getSaversWithdrawQuote(
+    asset: string,
+    address: string,
+    height?: number,
+    withdrawBps?: number,
+  ): Promise<QuoteSaverWithdrawResponse> {
+    for (const api of this.quoteApi) {
+      try {
+        const resp = (await api.quotesaverwithdraw(height, asset, address, withdrawBps)).data
+        return resp
+      } catch (e) {
+        //console.error(e)
+      }
+    }
+    throw new Error(`THORNode not responding`)
+  }
+
+  /**
+   *
+   * @param fromAsset - input asset
+   * @param toAsset - output asset
+   * @param amount - amount to swap
+   * @param destination - vault address
+   * @param toleranceBps - slip percent
+   * @param affiliateBps - affiliate percent
+   * @param affiliate - affiliate address
+   * @param height - block height
+   * @returns quotes swap object response
+   */
+  async getSwapQuote(
+    fromAsset: string,
+    toAsset: string,
+    amount: number,
+    destination: string,
+    toleranceBps: number,
+    affiliateBps: number,
+    affiliate: string,
+    height?: number,
+  ): Promise<QuoteSwapResponse> {
+    for (const api of this.quoteApi) {
+      try {
+        const resp = (
+          await api.quoteswap(height, fromAsset, toAsset, amount, destination, toleranceBps, affiliateBps, affiliate)
+        ).data
         return resp
       } catch (e) {
         //console.error(e)